+++ Source By Mr Blue/src/nro/models/boss/sieu_hang/Rival.java
@@ -38,7 +38,7 @@ public class Rival extends SuperRank {
                 player.gender,
                 new short[]{player.getHead(), player.getBody(), player.getLeg(), player.getFlagBag(), player.getAura(), player.getEffFront()},
                 player.nPoint.dameg,
-                new int[]{player.nPoint.hpg},
+                new int[]{(int)player.nPoint.hpg},
                 new int[]{113},
                 skillTemp,
                 new String[]{}, //text chat 1
